+++ replay_test.go
@@ -1,39 +1,38 @@
-package sse_test
+package sse
 
 import (
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"github.com/tmaxmax/go-sse"
 )
 
-func msg(tb testing.TB, data, id string, expiry time.Duration, topic string) *sse.Message {
+func msg(tb testing.TB, data, id, topic string) *Message {
 	tb.Helper()
 
-	e := &sse.Message{Topic: topic, ExpiresAt: time.Now().Add(expiry)}
+	e := &Message{Topic: topic}
 	e.AppendData(data)
 	if id != "" {
-		e.ID = sse.ID(id)
+		e.ID = ID(id)
 	}
 
 	return e
 }
 
-func replay(tb testing.TB, p sse.ReplayProvider, lastEventID sse.EventID, topics ...string) []*sse.Message {
+func replay(tb testing.TB, p ReplayProvider, lastEventID EventID, topics ...string) []*Message {
 	tb.Helper()
 
 	if len(topics) == 0 {
-		topics = []string{sse.DefaultTopic}
+		topics = []string{DefaultTopic}
 	}
 
-	var replayed []*sse.Message
-	cb := func(m *sse.Message) bool {
+	var replayed []*Message
+	cb := func(m *Message) bool {
 		replayed = append(replayed, m)
 		return true
 	}
 
-	sub := sse.Subscription{
+	sub := Subscription{
 		Callback:    cb,
 		LastEventID: lastEventID,
 		Topics:      topics,
@@ -41,38 +40,49 @@ func replay(tb testing.TB, p sse.ReplayProvider, lastEventID sse.EventID, topics
 
 	_ = p.Replay(sub)
 
-	sub.LastEventID = sse.EventID{}
+	sub.LastEventID = EventID{}
 	_ = p.Replay(sub)
 
-	sub.LastEventID = sse.ID("mama")
+	sub.LastEventID = ID("mama")
 	_ = p.Replay(sub)
 
-	sub.LastEventID = sse.ID("10")
+	sub.LastEventID = ID("10")
 	_ = p.Replay(sub)
 
 	return replayed
 }
 
-func testReplayError(tb testing.TB, p sse.ReplayProvider) {
+func testReplayError(tb testing.TB, p ReplayProvider) {
 	tb.Helper()
 
+	_, isExpiry := p.(*ValidReplayProvider)
+
+	var now time.Time
+	if isExpiry {
+		now = time.Now()
+		timeNow = func() time.Time { return now.Add(time.Hour) }
+	}
+
 	putMessages(p,
-		msg(tb, "a", "1", time.Hour, sse.DefaultTopic),
-		msg(tb, "b", "2", time.Hour, sse.DefaultTopic),
+		msg(tb, "a", "1", DefaultTopic),
+		msg(tb, "b", "2", DefaultTopic),
 	)
 
-	cb := func(_ *sse.Message) bool { return false }
+	cb := func(_ *Message) bool { return false }
+	if isExpiry {
+		timeNow = func() time.Time { return now }
+	}
 
-	success := p.Replay(sse.Subscription{
+	success := p.Replay(Subscription{
 		Callback:    cb,
-		LastEventID: sse.ID("1"),
-		Topics:      []string{sse.DefaultTopic},
+		LastEventID: ID("1"),
+		Topics:      []string{DefaultTopic},
 	})
 
 	require.False(tb, success, "received invalid error")
 }
 
-func putMessages(p sse.ReplayProvider, msgs ...*sse.Message) {
+func putMessages(p ReplayProvider, msgs ...*Message) {
 	for i := range msgs {
 		msgs[i] = p.Put(msgs[i])
 	}
@@ -80,58 +90,71 @@ func putMessages(p sse.ReplayProvider, msgs ...*sse.Message) {
 
 func TestValidReplayProvider(t *testing.T) {
 	t.Parallel()
+	t.Cleanup(func() { timeNow = time.Now })
 
-	p := sse.NewValidReplayProvider(true)
+	p := &ValidReplayProvider{
+		TTL:     time.Millisecond * 5,
+		AutoIDs: true,
+	}
 
 	require.NoError(t, p.GC(), "unexpected GC error") // no elements, noop
 
-	exp := time.Millisecond * 5
+	now := time.Now()
+	initialNow := now
+	timeNow = func() time.Time { return now }
 
 	putMessages(p,
-		msg(t, "hi", "", exp, sse.DefaultTopic),
-		msg(t, "there", "", exp, "t"),
-		msg(t, "world", "", exp*2, sse.DefaultTopic),
-		msg(t, "again", "", exp*2, "t"),
-		msg(t, "world", "", exp*5, sse.DefaultTopic),
-		msg(t, "x", "", exp*5, "t"),
-		msg(t, "again", "", exp*10, "t"),
+		msg(t, "hi", "", DefaultTopic),
+		msg(t, "there", "", "t"),
 	)
+	now = now.Add(p.TTL)
+	putMessages(p,
+		msg(t, "world", "", DefaultTopic),
+		msg(t, "again", "", "t"),
+	)
+	now = now.Add(p.TTL * 3)
+	putMessages(p,
+		msg(t, "world", "", DefaultTopic),
+		msg(t, "x", "", "t"),
+	)
+	now = now.Add(p.TTL * 5)
+	putMessages(p, msg(t, "again", "", "t"))
 
-	time.Sleep(exp)
+	now = initialNow.Add(p.TTL)
 
 	require.NoError(t, p.GC(), "unexpected GC error")
 
-	time.Sleep(exp)
+	now = now.Add(p.TTL)
 
-	replayed := replay(t, p, sse.ID("3"))[0]
+	replayed := replay(t, p, ID("3"))[0]
 	require.Equal(t, "id: 4\ndata: world\n\n", replayed.String())
 
-	testReplayError(t, sse.NewValidReplayProvider())
+	testReplayError(t, &ValidReplayProvider{})
 }
 
 func TestFiniteReplayProvider(t *testing.T) {
 	t.Parallel()
 
-	p := sse.NewFiniteReplayProvider(3)
+	p := NewFiniteReplayProvider(3)
 
 	putMessages(p,
-		msg(t, "", "1", 0, ""),
-		msg(t, "hello", "2", 0, sse.DefaultTopic),
-		msg(t, "there", "3", 0, "t"),
-		msg(t, "world", "4", 0, sse.DefaultTopic),
+		msg(t, "", "1", ""),
+		msg(t, "hello", "2", DefaultTopic),
+		msg(t, "there", "3", "t"),
+		msg(t, "world", "4", DefaultTopic),
 	)
 
-	replayed := replay(t, p, sse.ID("2"))[0]
+	replayed := replay(t, p, ID("2"))[0]
 	require.Equal(t, "id: 4\ndata: world\n\n", replayed.String())
 
 	putMessages(p,
-		msg(t, "", "5", 0, "t"),
-		msg(t, "", "6", 0, "t"),
-		msg(t, "again", "7", 0, sse.DefaultTopic),
+		msg(t, "", "5", "t"),
+		msg(t, "", "6", "t"),
+		msg(t, "again", "7", DefaultTopic),
 	)
 
-	replayed = replay(t, p, sse.ID("4"))[0]
+	replayed = replay(t, p, ID("4"))[0]
 	require.Equal(t, "id: 7\ndata: again\n\n", replayed.String())
 
-	testReplayError(t, sse.NewFiniteReplayProvider(10))
+	testReplayError(t, NewFiniteReplayProvider(10))
 }